[AUTO] fix: calendar tab — UTC date bug, week view, nav, sync settings, view toggle
Fixes 11 calendar tab findings from walkthrough:

- Fix 1 (utils.js): Add parseLocalDate() to parse YYYY-MM-DD as local midnight,
  preventing UTC off-by-one date rendering in negative-offset timezones
- Fix 2 (pages): selectedDate in day view parsed via parseLocalDate
- Fix 3 (pages): Split getWeekDays() into weekDays (hero strip, always today)
  and viewWeekDays (week view, anchored to selectedDate)
- Fix 4a/4b (pages): Week view title shows date range (Apr 6 – Apr 12, 2026);
  week view header and body use viewWeekDays
- Fix 5 (pages): "This Week" stat card label includes actual date range
- Fix 6 (pages): "Schedule Live Show" button renamed to "Whatnot Live"
- Fix 7 (handlers): navigateCalendar uses parseLocalDate instead of new Date
- Fix 8 (handlers): navigateCalendarMonth now also sets selectedCalendarDate
  to the first of the new month so right sidebar updates
- Fix 9 (handlers): Sync Settings removes internal env var names, shows
  user-friendly "contact support" messaging instead
- Fix 10 (widgets): calendarTimeline uses weekday:short to prevent wrapping
- Fix 11 (css): Active view toggle button gets border for visual clarity

Verified: node --check passes all 4 edited source files; both bundles
rebuilt successfully (core-bundle.js 1436KB, chunk-deferred.js 1941KB)

+++ src/frontend/ui/widgets.js
@@ -4947,7 +4947,7 @@ const calendarTimeline = {
         return `
             <div class="calendar-timeline">
                 <div class="calendar-timeline-header">
-                    <span>${date.toLocaleDateString('en-US', { weekday: 'long', month: 'short', day: 'numeric' })}</span>
+                    <span>${date.toLocaleDateString('en-US', { weekday: 'short', month: 'short', day: 'numeric' })}</span>
                     <span class="text-sm text-gray-500">${dayEvents.length} events</span>
                 </div>
                 ${dayEvents.length > 0 ? dayEvents.map(event => `
